+++ UnitedOperationsFramework/MapCompassRemoverModule/cfgFunctions.hpp
@@ -1,0 +1,10 @@
+class UO_FW_MapCompassRemoverFunctions
+{
+	file = "\x\UO_FW\addons\main\MapCompassRemoverModule\cfgFunctions\common";
+	class MapCompassRemoverInit
+	{
+		postInit = 1;
+	};
+	
+
+};