+++ CHANGELOG.md
@@ -8,8 +8,9 @@ All notable changes to this project will be documented in this file.
 ## [1.0.1] - 2018-11-11
 ### Added
 - Added Button to setup respawn and settings in 3den
+- Added Map Compass Remover
 ### Changed
-
+- Removed briefing limitations
 ### Fixed
 - Ai Driver wasn't being added when ace menued
 ### Removed